Use static generators
